Pass objects into CSVDSL#generate
CSVDSL#to_csv was removed and renamed to #generate.  Objects are no
longer passed into the initializer but rather to the generate method.
This addresses a problem where the objects were not being sorted for
CSVs in the same way they were for packing slips.  It is also much
cleaner.